pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "content-pilot"
version = "0.1.0"
description = "AI-driven social media automation tool for Chinese platforms"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Content Pilot Team" }]
keywords = ["social-media", "automation", "ai", "content-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "playwright>=1.40",
    "click>=8.1",
    "apscheduler>=3.10",
    "anthropic>=0.40",
    "openai>=1.10",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "qrcode[pil]>=7.4",
    "rich>=13.7",
    "aiosqlite>=0.19",
    "Pillow>=10.2",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.2",
]

[project.scripts]
content-pilot = "content_pilot.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
